Updated VCU118 | Bumped naming on Arty
